Cut demodulator and DSP CPU by 24% on a Pi 4, 15% on a Pi 5

+++ include/InputBufferReader.hpp
@@ -22,7 +22,7 @@ public:
           m_reader(ringBuffer)
     { }
 
-    inline void readMagnitude(float* out) {
+    inline void readMagnitude(int32_t* out) {
         m_reader.process([&](const RawType* buffer) {
             this->processBlock(buffer, out);
         });
